+++ pyccel/codegen/printing/ccode.py
@@ -2836,6 +2836,24 @@ class CCodePrinter(CodePrinter):
         dict_var = self._print(ObjectAddress(expr.dict_obj))
         return f"{c_type}_clear({dict_var});\n"
 
+    def _print_DictGet(self, expr):
+        target = expr.dict_obj
+        class_type = target.class_type
+        c_type = self.get_c_type(class_type)
+
+        dict_var = self._print(ObjectAddress(target))
+        key = self._print(expr.key)
+
+        if expr.default_value is not None:
+            default_val = self._print(expr.default_value)
+            return f"({c_type}_contains({dict_var}, {key}) ? *{c_type}_at({dict_var}, {key}) : {default_val})"
+
+        if target.is_optional:
+            return f"({c_type}_contains({dict_var}, {key}) ? *{c_type}_at({dict_var}, {key}) : NULL)"
+
+        errors.report("Accessing dictionary using get without default might cause runtime issues", severity="warning")
+        return f"*{c_type}_at({dict_var}, {key})"
+
     def _print_DictPop(self, expr):
         target = expr.dict_obj
         class_type = target.class_type
@@ -2844,7 +2862,7 @@ class CCodePrinter(CodePrinter):
         dict_var = self._print(ObjectAddress(target))
         key = self._print(expr.key)
 
-        if expr.default_value:
+        if expr.default_value is not None:
             default = self._print(expr.default_value)
             pop_expr = f"{c_type}_pop_with_default({dict_var}, {key}, {default})"
         else: